backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
PyMuPDF>=1.24.9
google-generativeai==0.8.3
groq==0.11.0
openai==1.51.0
anthropic==0.34.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
